feat: migrate from Deno libraries to node ones + vitest, and publish in npm
Remove the old deno setup and JSR promotion and replace with node and npm, to be compatible with the
old services. We add `dist` to the repo so that we can use directly from github in TS projects
without issues.

BREAKING CHANGE: Drop the JSR package and replace with NPM one. Replace @std/* with node lib.