fix: sort-keys peerDependency 4.2
sort-keys was added as a dependency, but it was added at the latest v5. That version is a pure ESM,
which makes this not usable for any application that needs to support CommonJS.

For now it will be locked at version 4, since depending on the usage in the clients, it can still end up getting 5.x and provoking errors.

BREAKING CHANGE: sort-keys as peerDependency, using v4.2